% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{glmlasso}
\alias{glmlasso}
\title{L1 penalized logistic regression and linear regression}
\usage{
glmlasso(X, y, family = c("gaussian", "binomial"), nlambda = 100,
  minlam = NULL, lambda = NULL, penalty.factor = c(0, rep(1, ncol(X) -
  1)), tol = 1e-06)
}
\arguments{
\item{X}{design matrix with first column being 1}

\item{y}{the response vector}

\item{family}{the distribution of the response. Currently only
support two type with "gaussian" doing the L1 linear regression
and "binomial" doing the L1 logistic regression}

\item{nlambda}{length of the lamda sequence with default value 100}

\item{minlam}{the minimun of the lambda sequence. By defalt, it is
lambdaMax/1000}

\item{lambda}{user specified lambda sequence. By default it is null
and the function will call logLamRange() or lmLamRange() to
calculate the lambda sequence}

\item{penalty.factor}{the vector containing the individual penalty
factor for each coefficient. By default, it is
c(0,rep(1,ncol(X)-1))}

\item{tol}{the precision of the estimated coefficients.}
}
\value{
an object of S3 class "glmlasso" containing the lambda
    sequence and an matrix for the solution path of coefficients
}
\description{
Use coordinate-desecent algorithm to provide the solution path for
L1 penalized logistic regression and linear regression
}
\examples{
ilogit <- function(x) exp(x)/(1+exp(x))
n <- 1000
p <- 10
Z <- matrix(rnorm(n*(p-2)),nrow=n,ncol=(p-2))
v2 <- c(rep(1,n/2),rep(0,n/2))
X <- cbind(1,v2,Z)
beta.star <- c(p^(-1/2)*rnorm(p-1),2)
y <- rbinom(n=n,size=1,prob=ilogit(X\%*\%beta.star))
glmlasso(X,y,family="binomial")
}
