% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lmLamRange}
\alias{lmLamRange}
\title{Calculate lambda sequence for L1 linear regression}
\usage{
lmLamRange(X, y, nlam, minlam, dweight)
}
\arguments{
\item{X}{design matrix with first column being 1}

\item{y}{response vector}

\item{nlam}{length of the lambda sequence}

\item{minlam}{smallest lambda for the lambda sequence}

\item{dweigtht}{vector of length ncol(X) with penalty factors for
each coefficient}
}
\value{
a list containing the largest lambda, the lambda sequence
and the estimated intercept at the largest lambda
}
\description{
This function calculate the smallest lamda that shrink all
coefficients but intercept to zero and return a sequence of lambda
used for glmlasso
}
\examples{
ilogit <- function(x) exp(x)/(1+exp(x))
n <- 1000
p <- 10
Z <- matrix(rnorm(n*(p-2)),nrow=n,ncol=(p-2))
v2 <- c(rep(1,n/2),rep(0,n/2))
X <- cbind(1,v2,Z)
beta.star <- c(p^(-1/2)*rnorm(p-1),2)
y <- rbinom(n=n,size=1,prob=ilogit(X\%*\%beta.star))
lmLamRange(X,y,100,0.0001,c(0,rep(1,p-1)))
}
