% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{plot.glmlasso}
\alias{plot.glmlasso}
\title{Plot the solution path glmlasso}
\usage{
\method{plot}{glmlasso}(x)
}
\arguments{
\item{x}{the S3 class "glmlasso" object given function glmlasso}
}
\value{
a plot with the Lasso solution path
}
\description{
Plot the solution path glmlasso
}
\examples{
ilogit <- function(x) exp(x)/(1+exp(x))
n <- 1000
p <- 10
Z <- matrix(rnorm(n*(p-2)),nrow=n,ncol=(p-2))
v2 <- c(rep(1,n/2),rep(0,n/2))
X <- cbind(1,v2,Z)
beta.star <- c(p^(-1/2)*rnorm(p-1),2)
y <- rbinom(n=n,size=1,prob=ilogit(X\%*\%beta.star))
out <- glmlasso(X,y,family="binomial")
plot(out)
}
